feat(tensor): Implement broadcasting for +
Implement broadcasting for the + operator to enable addition of two
tensors with different but broadcastable shapes by adding
`compute_strides()`, `pad_vector()`, `broadcast_result_shape()`,
`broadcast_strides()`, `compute_offset()`, and `next_index()`
functions. Add a new constructor for the Tensor class used by creation
methods that specify a shape as a vector, and a new version of
`empty()` that's compatible with this.