feat(functional): Add unfold_cuda() function
Add the unfold_cuda() function to extract sliding local blocks from a
batched input tensor that runs on the GPU using a custom CUDA kernel.
Add `float* device_data` to Tensor and `use_cuda` flag to the Tensor
constructor and empty() function that take a shape as an
initializer_list for use with unfold_cuda() and future CUDA kernels.