fix(tensor): Fix batch dims issue in matmul()
Fix an issue in matmul() where if A and B both have the same batch
dimensions which are greater than 1, result_batch_dims was not being
initialized. Fix an issue where if A or B have a single dimension, the
strides were not being updated based on the size 1 dim that was added.