perf(tensor): Fix redundant allocation in matmul
Remove redundant allocation of result.data with empty() in matmul() and
instead just set result.data to C for improved performance. Remove
cudaDeviceSynchronize() for better performance. Start testing the use
of a universal data shared_ptr for Tensors on the host or device by
adding cuda_fill() for creating Tensors on the GPU with ones().